rename: Icon~ 컴포넌트 폴더 세부 분리
