+++ src/components/Icon/index.js
@@ -1,9 +1,9 @@
-export { default as IconFacebook } from './IconFacebook';
-export { default as IconInstagram } from './IconInstagram';
-export { default as IconKebabMenu } from './IconKebabMenu';
-export { default as IconLink } from './IconLink';
-export { default as IconLogo } from './IconLogo';
-export { default as IconSearch } from './IconSearch';
-export { default as IconStar } from './IconStar';
-export { default as IconTwitter } from './IconTwitter';
-export { default as IconYoutube } from './IconYoutube';
+export { default as IconFacebook } from './svg/sns/IconFacebook';
+export { default as IconInstagram } from './svg/sns/IconInstagram';
+export { default as IconKebabMenu } from './svg/IconKebabMenu';
+export { default as IconLink } from './svg/IconLink';
+export { default as IconLogo } from './svg/IconLogo';
+export { default as IconSearch } from './svg/IconSearch';
+export { default as IconStar } from './svg/IconStar';
+export { default as IconTwitter } from './svg/sns/IconTwitter';
+export { default as IconYoutube } from './svg/sns/IconYoutube';
