+++ src/components/Icon/svg/IconLogo.jsx
