+++ src/components/Icon/svg/IconSearch.jsx
