+++ src/components/Icon/svg/IconStar.jsx
