+++ src/components/Icon/svg/sns/IconYoutube.jsx
